// src/interfaces/registrodeservicio.ts
export interface RegistrarServ {
    nombreConductor: string;
    apellidoConductor: string;
    tipoServicios: string;
    tiposemiremolque: string;
    patentetracto: string;
    patentesemirremolque: string;
    numeroguia: string;
    informacionAdi: string;
    fechapresentacion: Date;
    horapresentacion: Date;
    horaretiro: Date;
    horadescarga: Date;
}